refactor: dataclass and file naming for queue info (#301)
* refactor: dataclass and file naming for queue info

* apply suggestions
