+++ src/braket/aws/aws_quantum_task.py
@@ -24,6 +24,7 @@ import boto3
 from braket.ahs.analog_hamiltonian_simulation import AnalogHamiltonianSimulation
 from braket.annealing.problem import Problem
 from braket.aws.aws_session import AwsSession
+from braket.aws.queue_information import QuantumTaskQueueInfo, QueuePriority
 from braket.circuits import Instruction
 from braket.circuits.circuit import Circuit, Gate, QubitSet
 from braket.circuits.circuit_helpers import validate_circuit_and_shots
@@ -54,7 +55,6 @@ from braket.error_mitigation import ErrorMitigation
 from braket.ir.blackbird import Program as BlackbirdProgram
 from braket.ir.openqasm import Program as OpenQASMProgram
 from braket.pulse.pulse_sequence import PulseSequence
-from braket.queue_information.queue_position import QueuePosition, QueuePriority
 from braket.schema_common import BraketSchemaBase
 from braket.task_result import (
     AnalogHamiltonianSimulationTaskResult,
@@ -315,12 +315,12 @@ class AwsQuantumTask(QuantumTask):
         """
         return self._status(use_cached_value)
 
-    def queue_position(self) -> QueuePosition:
+    def queue_position(self) -> QuantumTaskQueueInfo:
         """
         The queue position details for the quantum task.
 
         Returns:
-            QueuePosition: Instance containing the queue position information
+            QuantumTaskQueueInfo: Instance containing the queue position information
             for the task.
 
             Note: The queue_position is only returned when task is not in
@@ -329,12 +329,12 @@ class AwsQuantumTask(QuantumTask):
         Examples:
             task status = QUEUED
             >>> task.queue_position()
-            QueuePosition(queue_position='2', queue_priority=<QueuePriority.NORMAL: 'Normal'>,
-            message=None)
+            QuantumTaskQueueInfo(queue_position='2',
+            queue_priority=<QueuePriority.NORMAL: 'Normal'>, message=None)
 
             task status = COMPLETED
             >>> task.queue_position()
-            QueuePosition(queue_position='2',
+            QuantumTaskQueueInfo(queue_position='2',
             queue_priority=<QueuePriority.NORMAL: 'Normal'>,
             message='Task is in COMPLETED status. AmazonBraket does
                         not show queue position for this status.')
@@ -345,9 +345,9 @@ class AwsQuantumTask(QuantumTask):
 
         if queue_position == "None":
             message = response["message"]
-            return QueuePosition(queue_position, queue_priority, message)
+            return QuantumTaskQueueInfo(queue_position, queue_priority, message)
 
-        return QueuePosition(queue_position, queue_priority)
+        return QuantumTaskQueueInfo(queue_position, queue_priority)
 
     def _status(self, use_cached_value: bool = False) -> str:
         metadata = self.metadata(use_cached_value)
